Move testapp code around so it is easier to add a plugin.
