+++ composer.json
@@ -9,7 +9,7 @@
 	],
 	"type": "cakephp-plugin",
 	"require": {
-		"cakephp/core": "~3.4"
+		"cakephp/core": "~3.5"
 	},
 	"require-dev": {
 		"cakephp/cakephp": "~3.5",
@@ -20,24 +20,22 @@
 	"minimum-stability": "stable",
 	"autoload": {
 		"psr-4": {
-			"Authorization\\": "src"
+			"Authorization\\": "src/"
 		}
 	},
 	"autoload-dev": {
 		"psr-4": {
-			"AuthTest\\": "tests/test_app/Plugin/AuthTest/src",
-			"Authorization\\Test\\": "tests",
-			"Authorization\\Test\\App\\": "tests/test_app/App",
-			"Cake\\Test\\": "./vendor/cakephp/cakephp/tests",
-			"TestApp\\": "tests/test_app",
-			"TestPlugin\\": "tests/test_app/Plugin/TestPlugin/src"
+			"AuthTest\\": "tests/test_app/Plugin/AuthTest/src/",
+			"Authorization\\Test\\": "tests/",
+			"Cake\\Test\\": "./vendor/cakephp/cakephp/tests/",
+			"TestApp\\": "tests/test_app/TestApp/",
+			"TestPlugin\\": "tests/test_app/Plugin/TestPlugin/src/"
 		}
 	},
 	"authors": [
-		{
-			"name": "Florian Krämer",
-			"homepage": "http://florian-kraemer.net",
-			"role": "Developer"
-		}
+			{
+					"name": "CakePHP Community",
+					"homepage": "https://github.com/cakephp/cakephp/graphs/contributors"
+			}
 	]
 }
