+++ tests/test_app/TestApp/Policy/ArticlePolicy.php
@@ -1,5 +1,5 @@
 <?php
-namespace TestApp\Authorization\Policy;
+namespace TestApp\Policy;
 
 class ArticlePolicy
 {
